ref(replay): extract formatting helpers and address review findings
- Extract replay formatting logic from replay/view.ts (950 lines) into
  a dedicated src/lib/formatters/replay.ts module, bringing view.ts down
  to ~425 lines focused on argument parsing and API orchestration.
- Replace this.stderr.write() with structured logger (consola) per repo
  conventions. Add diagnostic logging to silent catch blocks.
- Document normalizeReplayProjectId coercion and empty-string project
  slug in org-wide issue search.
- Add explanatory comment for parseSingleArg's org/hex-id special case.
